+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Made feature a property of the like button.
 - Changed feature list to be source-based.
 - Added easeTo and flyTo methods to gl-map.
+- Removed gl-property and gl-template components.
 
 # [0.2.1] - 2018-06-01
 - Ensure that gl-form-page is loaded by dependent modules.